add default templates
templates with key "default: true" are automatically added to the top of
new shift reports.

together with "required" fields (added in
dd430465a5ec37f39af3e6ab6e511009b80b3c80), which cannot be deleted from
the template, we can pretty strongly hint to the user what they should
be filling out.

for now, that's as strict as we'll be, since any such enforcement is a
departure from the (sometimes useful!) manhattan field free-for-all.