save on interval, only if changes were made
with nontrivial latency, the per-character document updates would race
and all fail. it has to grab the current revision to update, and 3 fast
updates with a high GET latency go:

    GET1 GET2 GET3 POST1 POST2 POST3

so only the first one has a correct _rev and worked.

now on character changes, clicks, etc., the editor is alerted that
the block needs to be saved. when the (5 second) interval timer clicks,
pending changes are saved.

as a bonus we end up with a lot fewer extraneous revisions and need less
database compaction.