improve conditions with multiple editors
Give each editor session a unique ID, and keep track of who last touched each
block. When someone is editing a block, it is disabled on other clients. If you
try to edit a report that you didn't create, everything is disabled until you
click a button to take over control (and lock out everyone else). If someone
else submits a report you're working on, you get alerted and booted back to the
index.

This makes it possible for multiple clients to edit a report, as long as
everyone is working on different blocks.